+++ examples/resources/iosxr_l2vpn_xconnect_group_p2p_interface/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_l2vpn_xconnect_group_p2p_interface.example "Cisco-IOS-XR-um-l2vpn-cfg:/l2vpn/xconnect/groups/group[group-name=P2P]/p2ps/p2p[p2p-xconnect-name=XC]/interfaces/interface[interface-name=GigabitEthernet0/0/0/2]"