+++ examples/resources/iosxr_l2vpn_xconnect_group_p2p_neighbor_ipv4/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_l2vpn_xconnect_group_p2p_neighbor_ipv4.example "Cisco-IOS-XR-um-l2vpn-cfg:/l2vpn/xconnect/groups/group[group-name=P2P]/p2ps/p2p[p2p-xconnect-name=XC]/neighbor/ipv4s/ipv4[address=2.3.4.5][pw-id=1]"