+++ examples/resources/iosxr_l2vpn_xconnect_group_p2p_neighbor_ipv6/import.sh
@@ -1,1 +1,0 @@
-terraform import iosxr_l2vpn_xconnect_group_p2p_neighbor_ipv6.example "Cisco-IOS-XR-um-l2vpn-cfg:/l2vpn/xconnect/groups/group[group-name=P2P]/p2ps/p2p[p2p-xconnect-name=XC]/neighbor/ipv6s/ipv6[address=2001::2][pw-id=2]"