+++ internal/provider/resource_iosxr_l2vpn_xconnect_group_p2p_test.go
@@ -18,6 +18,10 @@ func TestAccIosxrL2VPNXconnectGroupP2P(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr("iosxr_l2vpn_xconnect_group_p2p.test", "p2p_xconnect_name", "XC"),
 					resource.TestCheckResourceAttr("iosxr_l2vpn_xconnect_group_p2p.test", "description", "My P2P Description"),
+					resource.TestCheckResourceAttr("iosxr_l2vpn_xconnect_group_p2p.test", "interfaces.0.interface_name", "GigabitEthernet0/0/0/2"),
+					resource.TestCheckResourceAttr("iosxr_l2vpn_xconnect_group_p2p.test", "ipv4_neighbors.0.address", "2.3.4.5"),
+					resource.TestCheckResourceAttr("iosxr_l2vpn_xconnect_group_p2p.test", "ipv4_neighbors.0.pw_id", "1"),
+					resource.TestCheckResourceAttr("iosxr_l2vpn_xconnect_group_p2p.test", "ipv4_neighbors.0.pw_class", "PW_CLASS_1"),
 				),
 			},
 			{
@@ -44,6 +48,14 @@ func testAccIosxrL2VPNXconnectGroupP2PConfig_all() string {
 		group_name = "P2P"
 		p2p_xconnect_name = "XC"
 		description = "My P2P Description"
+		interfaces = [{
+			interface_name = "GigabitEthernet0/0/0/2"
+		}]
+		ipv4_neighbors = [{
+			address = "2.3.4.5"
+			pw_id = 1
+			pw_class = "PW_CLASS_1"
+		}]
 	}
 	`
 }